--- src/constants/actionTypes.js
@@ -1,3 +1,0 @@
-export const REQUEST_USER = 'REQUEST_USER';
-export const RECEIVE_USER = 'RECEIVE_USER';
-export const SET_LOGIN_USER = 'SET_LOGIN_USER';
